requirements.txt
fastapi
uvicorn
aiohttp
pydantic
beautifulsoup4
pytest
pytest-asyncio
redis
python-dotenv